removed old src/dcp/common.c and .h and reliance on global mpi rank
