+++ src/Makefile.am
@@ -1,1 +1,1 @@
-SUBDIRS = common dbcast dcmp dchmod dcp1 dcp ddup dfilemaker drm dstripe dwalk
+SUBDIRS = common dbcast dcmp dchmod dcp ddup dfilemaker drm dstripe dwalk
