+++ src/common/Makefile.am
@@ -16,7 +16,6 @@ include_HEADERS = \
     mfu_param_path.h \
     mfu_path.h \
     mfu_util.h \
-    common.h \
     handle_args.h 
 
 libmfu_la_SOURCES = \
@@ -32,7 +31,6 @@ libmfu_la_SOURCES = \
     mfu_path.c \
     mfu_util.c \
     strmap.c \
-    common.c \
     handle_args.c 
 
 libmfu_la_CPPFLAGS = \
