+++ src/common/mfu_flist.c
@@ -65,6 +65,15 @@
 #include "strmap.h"
 #include "handle_args.h"
 
+/** Where we should keep statistics related to this file copy. */
+DCOPY_statistics_t DCOPY_statistics;
+
+/** Options specified by the user. */
+DCOPY_options_t DCOPY_user_opts;
+
+mfu_copy_file_cache_t mfu_copy_src_cache;
+mfu_copy_file_cache_t mfu_copy_dst_cache;
+
 /****************************************
  * Functions on types
  ***************************************/
@@ -457,6 +466,427 @@ void mfu_flist_free(mfu_flist* pbflist)
     return;
 }
 
+static int DCOPY_open_file(const char* file, int read_flag, mfu_copy_file_cache_t* cache)
+{
+    int newfd = -1;
+
+    /* see if we have a cached file descriptor */
+    char* name = cache->name;
+    if (name != NULL) {
+        /* we have a cached file descriptor */
+        int fd = cache->fd;
+        if (strcmp(name, file) == 0 && cache->read == read_flag) {
+            /* the file we're trying to open matches name and read/write mode,
+             * so just return the cached descriptor */
+            return fd;
+        } else {
+            /* the file we're trying to open is different,
+             * close the old file and delete the name */
+            mfu_close(name, fd);
+            mfu_free(&cache->name);
+        }
+    }
+
+    /* open the new file */
+    if (read_flag) {
+        int flags = O_RDONLY;
+        if (DCOPY_user_opts.synchronous) {
+            flags |= O_DIRECT;
+        }
+        newfd = mfu_open(file, flags);
+    } else {
+        int flags = O_WRONLY | O_CREAT;
+        if (DCOPY_user_opts.synchronous) {
+            flags |= O_DIRECT;
+        }
+        newfd = mfu_open(file, flags, DCOPY_DEF_PERMS_FILE);
+    }
+
+    /* cache the file descriptor */
+    if (newfd != -1) {
+        cache->name = MFU_STRDUP(file);
+        cache->fd   = newfd;
+        cache->read = read_flag;
+#ifdef LUSTRE_SUPPORT
+        /* Zero is an invalid ID for grouplock. */
+        if (DCOPY_user_opts.grouplock_id != 0) {
+            int rc;
+
+            rc = ioctl(newfd, LL_IOC_GROUP_LOCK, DCOPY_user_opts.grouplock_id);
+            if (rc) {
+                MFU_LOG(MFU_LOG_ERR, "Failed to obtain grouplock with ID %d "
+                    "on file `%s', ignoring this error: %s",
+                    DCOPY_user_opts.grouplock_id,
+                    file, strerror(errno));
+            } else {
+                MFU_LOG(MFU_LOG_INFO, "Obtained grouplock with ID %d "
+                    "on file `%s', fd %d", DCOPY_user_opts.grouplock_id,
+                    file, newfd);
+            }
+        }
+#endif
+    }
+
+    return newfd;
+}
+
+/* copy all extended attributes from op->operand to dest_path */
+static void DCOPY_copy_xattrs(
+    mfu_flist flist,
+    uint64_t idx,
+    const char* dest_path)
+{
+#if DCOPY_USE_XATTRS
+    /* get source file name */
+    const char* src_path = mfu_flist_file_get_name(flist, idx);
+
+    /* start with a reasonable buffer, we'll allocate more as needed */
+    size_t list_bufsize = 1204;
+    char* list = (char*) MFU_MALLOC(list_bufsize);
+
+    /* get list, if list_size == ERANGE, try again */
+    ssize_t list_size;
+    int got_list = 0;
+
+    /* get current estimate for list size */
+    while(! got_list) {
+        list_size = llistxattr(src_path, list, list_bufsize);
+
+        if(list_size < 0) {
+            if(errno == ERANGE) {
+                /* buffer is too small, free our current buffer
+                 * and call it again with size==0 to get new size */
+                mfu_free(&list);
+                list_bufsize = 0;
+            }
+            else if(errno == ENOTSUP) {
+                /* this is common enough that we silently ignore it */
+                break;
+            }
+            else {
+                /* this is a real error */
+                MFU_LOG(MFU_LOG_ERR, "Failed to get list of extended attributes on %s llistxattr() errno=%d %s",
+                    src_path, errno, strerror(errno)
+                   );
+                break;
+            }
+        }
+        else {
+            if(list_size > 0 && list_bufsize == 0) {
+                /* called llistxattr with size==0 and got back positive
+                 * number indicating size of buffer we need to allocate */
+                list_bufsize = (size_t) list_size;
+                list = (char*) MFU_MALLOC(list_bufsize);
+            }
+            else {
+                /* got our list, it's size is in list_size, which may be 0 */
+                got_list = 1;
+            }
+        }
+    }
+
+    /* iterate over list and copy values to new object lgetxattr/lsetxattr */
+    if(got_list) {
+        char* name = list;
+
+        while(name < list + list_size) {
+            /* start with a reasonable buffer,
+             * allocate something bigger as needed */
+            size_t val_bufsize = 1024;
+            void* val = (void*) MFU_MALLOC(val_bufsize);
+
+            /* lookup value for name */
+            ssize_t val_size;
+            int got_val = 0;
+
+            while(! got_val) {
+                val_size = lgetxattr(src_path, name, val, val_bufsize);
+
+                if(val_size < 0) {
+                    if(errno == ERANGE) {
+                        /* buffer is too small, free our current buffer
+                         * and call it again with size==0 to get new size */
+                        mfu_free(&val);
+                        val_bufsize = 0;
+                    }
+                    else if(errno == ENOATTR) {
+                        /* source object no longer has this attribute,
+                         * maybe deleted out from under us */
+                        break;
+                    }
+                    else {
+                        /* this is a real error */
+                        MFU_LOG(MFU_LOG_ERR, "Failed to get value for name=%s on %s llistxattr() errno=%d %s",
+                            name, src_path, errno, strerror(errno)
+                           );
+                        break;
+                    }
+                }
+                else {
+                    if(val_size > 0 && val_bufsize == 0) {
+                        /* called lgetxattr with size==0 and got back positive
+                         * number indicating size of buffer we need to allocate */
+                        val_bufsize = (size_t) val_size;
+                        val = (void*) MFU_MALLOC(val_bufsize);
+                    }
+                    else {
+                        /* got our value, it's size is in val_size, which may be 0 */
+                        got_val = 1;
+                    }
+                }
+            }
+
+            /* set attribute on destination object */
+            if(got_val) {
+                int setrc = lsetxattr(dest_path, name, val, (size_t) val_size, 0);
+
+                if(setrc != 0) {
+                    MFU_LOG(MFU_LOG_ERR, "Failed to set value for name=%s on %s llistxattr() errno=%d %s",
+                        name, dest_path, errno, strerror(errno)
+                       );
+                }
+            }
+
+            /* free value string */
+            mfu_free(&val);
+            val_bufsize = 0;
+
+            /* jump to next name */
+            size_t namelen = strlen(name) + 1;
+            name += namelen;
+        }
+    }
+
+    /* free space allocated for list */
+    mfu_free(&list);
+    list_bufsize = 0;
+
+    return;
+#endif /* DCOPY_USE_XATTR */
+}
+
+static void DCOPY_copy_ownership(
+    mfu_flist flist,
+    uint64_t idx,
+    const char* dest_path)
+{
+    /* get user id and group id of file */
+    uid_t uid = (uid_t) mfu_flist_file_get_uid(flist, idx);
+    gid_t gid = (gid_t) mfu_flist_file_get_gid(flist, idx);
+
+    /* note that we use lchown to change ownership of link itself, it path happens to be a link */
+    if(mfu_lchown(dest_path, uid, gid) != 0) {
+        /* TODO: are there other EPERM conditions we do want to report? */
+
+        /* since the user running dcp may not be the owner of the
+         * file, we could hit an EPERM error here, and the file
+         * will be left with the effective uid and gid of the dcp
+         * process, don't bother reporting an error for that case */
+        if (errno != EPERM) {
+            MFU_LOG(MFU_LOG_ERR, "Failed to change ownership on %s lchown() errno=%d %s",
+                dest_path, errno, strerror(errno)
+               );
+        }
+    }
+
+    return;
+}
+
+/* TODO: condionally set setuid and setgid bits? */
+static void DCOPY_copy_permissions(
+    mfu_flist flist,
+    uint64_t idx,
+    const char* dest_path)
+{
+    /* get mode and type */
+    mfu_filetype type = mfu_flist_file_get_type(flist, idx);
+    mode_t mode = (mode_t) mfu_flist_file_get_mode(flist, idx);
+
+    /* change mode */
+    if(type != MFU_TYPE_LINK) {
+        if(mfu_chmod(dest_path, mode) != 0) {
+            MFU_LOG(MFU_LOG_ERR, "Failed to change permissions on %s chmod() errno=%d %s",
+                dest_path, errno, strerror(errno)
+               );
+        }
+    }
+
+    return;
+}
+
+static void DCOPY_copy_timestamps(
+    mfu_flist flist,
+    uint64_t idx,
+    const char* dest_path)
+{
+    /* get atime seconds and nsecs */
+    uint64_t atime      = mfu_flist_file_get_atime(flist, idx);
+    uint64_t atime_nsec = mfu_flist_file_get_atime_nsec(flist, idx);
+
+    /* get mtime seconds and nsecs */
+    uint64_t mtime      = mfu_flist_file_get_mtime(flist, idx);
+    uint64_t mtime_nsec = mfu_flist_file_get_mtime_nsec(flist, idx);
+
+    /* fill in time structures */
+    struct timespec times[2];
+    times[0].tv_sec  = (time_t) atime;
+    times[0].tv_nsec = (long)   atime_nsec;
+    times[1].tv_sec  = (time_t) mtime;
+    times[1].tv_nsec = (long)   mtime_nsec;
+
+    /* set times with nanosecond precision using utimensat,
+     * assume path is relative to current working directory,
+     * if it's not absolute, and set times on link (not target file)
+     * if dest_path refers to a link */
+    if(utimensat(AT_FDCWD, dest_path, times, AT_SYMLINK_NOFOLLOW) != 0) {
+        MFU_LOG(MFU_LOG_ERR, "Failed to change timestamps on %s utime() errno=%d %s",
+            dest_path, errno, strerror(errno)
+           );
+    }
+
+#if 0
+    /* TODO: see stat-time.h and get_stat_atime/mtime/ctime to read sub-second times,
+     * and use utimensat to set sub-second times */
+    /* as last step, change timestamps */
+    if(! S_ISLNK(statbuf->st_mode)) {
+        struct utimbuf times;
+        times.actime  = statbuf->st_atime;
+        times.modtime = statbuf->st_mtime;
+        if(utime(dest_path, &times) != 0) {
+            MFU_LOG(MFU_LOG_ERR, "Failed to change timestamps on %s utime() errno=%d %s",
+                dest_path, errno, strerror(errno)
+               );
+        }
+    }
+    else {
+        struct timeval tv[2];
+        tv[0].tv_sec  = statbuf->st_atime;
+        tv[0].tv_usec = 0;
+        tv[1].tv_sec  = statbuf->st_mtime;
+        tv[1].tv_usec = 0;
+        if(lutimes(dest_path, tv) != 0) {
+            MFU_LOG(MFU_LOG_ERR, "Failed to change timestamps on %s utime() errno=%d %s",
+                dest_path, errno, strerror(errno)
+               );
+        }
+    }
+#endif
+
+    return;
+}
+
+/* called by single process upon detection of a problem */
+void DCOPY_abort(int code)
+{
+    MPI_Abort(MPI_COMM_WORLD, code);
+    exit(code);
+}
+
+/* called globally by all procs to exit */
+void DCOPY_exit(int code)
+{
+    /* CIRCLE_finalize or will this hang? */
+    mfu_finalize();
+    MPI_Finalize();
+    exit(code);
+}
+
+int mfu_copy_close_file(mfu_copy_file_cache_t* cache)
+{
+    int rc = 0;
+
+    /* close file if we have one */
+    char* name = cache->name;
+    if (name != NULL) {
+        int fd = cache->fd;
+
+        /* if open for write, fsync */
+        int read_flag = cache->read;
+        if (! read_flag) {
+            rc = mfu_fsync(name, fd);
+        }
+
+        /* close the file and delete the name string */
+        rc = mfu_close(name, fd);
+        mfu_free(&cache->name);
+    }
+
+    return rc;
+}
+
+/* iterate through list of files and set ownership, timestamps,
+ * and permissions starting from deepest level and working upwards,
+ * we go in this direction in case updating a file updates its
+ * parent directory */
+void mfu_copy_set_metadata(int levels, int minlevel, mfu_flist* lists)
+{
+    /* get current rank */
+    int rank;
+    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+
+    if (rank == 0) {
+        if(DCOPY_user_opts.preserve) {
+            MFU_LOG(MFU_LOG_INFO, "Setting ownership, permissions, and timestamps.");
+        }
+        else {
+            MFU_LOG(MFU_LOG_INFO, "Fixing permissions.");
+        }
+    }
+
+    /* now set timestamps on files starting from deepest level */
+    int level;
+    for (level = levels-1; level >= 0; level--) {
+        /* get list at this level */
+        mfu_flist list = lists[level];
+
+        /* cycle through our list of items and set timestamps
+         * for each one at this level */
+        uint64_t idx;
+        uint64_t size = mfu_flist_size(list);
+        for (idx = 0; idx < size; idx++) {
+            /* get type of item */
+            mfu_filetype type = mfu_flist_file_get_type(list, idx);
+
+            /* we've already set these properties for links,
+             * so we can skip those here */
+            if (type == MFU_TYPE_LINK) {
+                continue;
+            }
+
+            /* TODO: skip file if it's not readable */
+
+            /* get destination name of item */
+            const char* name = mfu_flist_file_get_name(list, idx);
+            char* dest = DCOPY_build_dest(name);
+
+            /* No need to copy it */
+            if (dest == NULL) {
+                continue;
+            }
+
+            if(DCOPY_user_opts.preserve) {
+                DCOPY_copy_ownership(list, idx, dest);
+                DCOPY_copy_permissions(list, idx, dest);
+                DCOPY_copy_timestamps(list, idx, dest);
+            }
+            else {
+                /* TODO: set permissions based on source permissons
+                 * masked by umask */
+                DCOPY_copy_permissions(list, idx, dest);
+            }
+
+            /* free destination item */
+            mfu_free(&dest);
+        }
+        
+        /* wait for all procs to finish before we start
+         * with files at next level */
+        MPI_Barrier(MPI_COMM_WORLD);
+    }
+
+    return;
+}
+
 static int mfu_create_directory(mfu_flist list, uint64_t idx)
 {
     /* get name of directory */
@@ -502,18 +932,21 @@ static int mfu_create_directory(mfu_flist list, uint64_t idx)
  * a child directory until the parent exists */
 int mfu_create_directories(int levels, int minlevel, mfu_flist* lists)
 {
+    /* get current rank */
+    int rank;
+    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+
     int rc = 0;
 
     int verbose = (mfu_debug_level <= MFU_LOG_INFO);
 
     /* indicate to user what phase we're in */
-    if (DCOPY_global_rank == 0) {
+    if (rank == 0) {
         MFU_LOG(MFU_LOG_INFO, "Creating directories.");
     }
 
     /* get our rank and number of ranks in job */
-    int rank, ranks;
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+    int ranks;
     MPI_Comm_size(MPI_COMM_WORLD, &ranks);
 
     /* work from shallowest level to deepest level */
@@ -710,18 +1143,21 @@ static int mfu_create_file(mfu_flist list, uint64_t idx)
 
 int mfu_create_files(int levels, int minlevel, mfu_flist* lists)
 {
+    /* get current rank */
+    int rank;
+    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+
     int rc = 0;
 
     int verbose = (mfu_debug_level <= MFU_LOG_INFO);
 
     /* indicate to user what phase we're in */
-    if (DCOPY_global_rank == 0) {
+    if (rank == 0) {
         MFU_LOG(MFU_LOG_INFO, "Creating files.");
     }
 
     /* get our rank and number of ranks in job */
-    int rank, ranks;
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+    int ranks;
     MPI_Comm_size(MPI_COMM_WORLD, &ranks);
 
     /* TODO: we don't need to have a barrier between levels */
@@ -1175,7 +1611,7 @@ static int mfu_copy_file(
     bool normal_copy_required;
 
     /* open the input file */
-    int in_fd = DCOPY_open_file(src, 1, &DCOPY_src_cache);
+    int in_fd = DCOPY_open_file(src, 1, &mfu_copy_src_cache);
     if (in_fd < 0) {
         MFU_LOG(MFU_LOG_ERR, "Failed to open input file `%s' errno=%d %s",
             src, errno, strerror(errno));
@@ -1183,7 +1619,7 @@ static int mfu_copy_file(
     }
 
     /* open the output file */
-    int out_fd = DCOPY_open_file(dest, 0, &DCOPY_dst_cache);
+    int out_fd = DCOPY_open_file(dest, 0, &mfu_copy_dst_cache);
     if (out_fd < 0) {
         MFU_LOG(MFU_LOG_ERR, "Failed to open output file `%s' errno=%d %s",
             dest, errno, strerror(errno));
@@ -1204,16 +1640,17 @@ static int mfu_copy_file(
 /* After receiving all incoming chunks, process open and write their chunks 
  * to the files. The process which writes the last chunk to each file also 
  * truncates the file to correct size.  A 0-byte file still has one chunk. */
-void mfu_copy_files(mfu_flist list)
+void mfu_copy_files(mfu_flist list, uint64_t chunk_size)
 {
+    /* get current rank */
+    int rank;
+    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+
     /* indicate which phase we're in to user */
-    if (DCOPY_global_rank == 0) {
+    if (rank == 0) {
         MFU_LOG(MFU_LOG_INFO, "Copying data.");
     }
 
-    /* get chunk size for copying files */
-    uint64_t chunk_size = (uint64_t)DCOPY_CHUNK_SIZE;
-    
     /* split file list into a linked list of file sections,
      * this evenly spreads the file sections across processes */
     mfu_file_chunk* p = mfu_file_chunk_list_alloc(list, chunk_size);
