+++ src/common/mfu_flist.h
@@ -29,12 +29,71 @@ extern "C" {
 #ifndef MFU_FLIST_H
 #define MFU_FLIST_H
 
+/* Make sure we're using 64 bit file handling. */
+#ifdef _FILE_OFFSET_BITS
+#undef _FILE_OFFSET_BITS
+#define _FILE_OFFSET_BITS 64
+#endif
+
+#ifndef _LARGEFILE64_SOURCE
+#define _LARGEFILE64_SOURCE 1
+#endif
+
+#ifndef __USE_LARGEFILE64
+#define __USE_LARGEFILE64
+#endif
+
+#ifndef _LARGEFILE_SOURCE
+#define _LARGEFILE_SOURCE
+#endif
+
+/* Enable posix extensions (popen). */
+#ifndef _BSD_SOURCE
+#define _BSD_SOURCE 1
+#endif
+
+/* For O_NOATIME support */
+#ifndef _GNU_SOURCE
+#define _GNU_SOURCE
+#endif
+
 #include <stdlib.h>
 #include <stdarg.h>
 #include <stdint.h>
 #include <ctype.h>
 #include <stdbool.h>
 #include "mpi.h"
+
+#if DCOPY_USE_XATTRS
+#include <attr/xattr.h>
+#endif /* DCOPY_USE_XATTRS */
+
+/* default mode to create new files or directories */
+#define DCOPY_DEF_PERMS_FILE (S_IRUSR | S_IWUSR)
+#define DCOPY_DEF_PERMS_DIR  (S_IRWXU)
+
+/* buffer size to read/write data to file system */
+#define FD_BLOCK_SIZE (1*1024*1024)
+
+/*
+ * FIXME: Is this description correct?
+ *
+ * This is the size of the buffer used to copy from the fd page cache to L1
+ * cache before the buffer is copied back down into the destination fd page
+ * cache.
+ */
+#define FD_PAGE_CACHE_SIZE (32768)
+
+#ifndef PATH_MAX
+#define PATH_MAX (4096)
+#endif
+
+#ifndef _POSIX_ARG_MAX
+#define MAX_ARGS 4096
+#else
+#define MAX_ARGS _POSIX_ARG_MAX
+#endif
+
 
 /****************************************
  * Define types
@@ -48,6 +107,55 @@ typedef enum mfu_filetypes_e {
     MFU_TYPE_DIR     = 3,
     MFU_TYPE_LINK    = 4,
 } mfu_filetype;
+
+typedef struct {
+    int64_t  total_dirs;         /* sum of all directories */
+    int64_t  total_files;        /* sum of all files */
+    int64_t  total_links;        /* sum of all symlinks */
+    int64_t  total_size;         /* sum of all file sizes */
+    int64_t  total_bytes_copied; /* total bytes written */
+    time_t   time_started;       /* time when dcp command started */
+    time_t   time_ended;         /* time when dcp command ended */
+    double   wtime_started;      /* time when dcp command started */
+    double   wtime_ended;        /* time when dcp command ended */
+} DCOPY_statistics_t;
+
+typedef struct {
+    int    copy_into_dir; /* flag indicating whether copying into existing dir */
+    char*  dest_path;     /* prefex of destination directory */
+    char*  input_file;    /* file name of input list*/
+    bool   preserve;      /* wether to preserve timestamps, ownership, permissions, etc. */
+    bool   synchronous;   /* whether to use O_DIRECT */
+    bool   sparse;        /* whether to create sparse files */
+    size_t chunk_size;    /* size to chunk files by */
+    size_t block_size;    /* block size to read/write to file system */
+    char*  block_buf1;    /* buffer to read / write data */
+    char*  block_buf2;    /* another buffer to read / write data */
+#ifdef LUSTRE_SUPPORT
+    int    grouplock_id;      /* Lustre grouplock ID */
+#endif
+} DCOPY_options_t;
+
+/* cache open file descriptor to avoid
+ * opening / closing the same file */
+typedef struct {
+    char* name; /* name of open file (NULL if none) */
+    int   read; /* whether file is open for read-only (1) or write (0) */
+    int   fd;   /* file descriptor */
+} mfu_copy_file_cache_t;
+
+/** Cache most recent open file descriptors. */
+//mfu_copy_file_cache_t mfu_copy_file_cache;
+
+/** Cache most recent open file descriptor to avoid opening / closing the same file */
+extern mfu_copy_file_cache_t mfu_copy_src_cache;
+extern mfu_copy_file_cache_t mfu_copy_dst_cache;
+
+/** Where we should keep statistics related to this file copy. */
+extern DCOPY_statistics_t DCOPY_statistics;
+
+/** Options specified by the user. */
+extern DCOPY_options_t DCOPY_user_opts;
 
 /* define handle type to a file list */
 typedef void* mfu_flist;
@@ -136,6 +244,50 @@ void mfu_flist_array_by_depth(
     int* outmin,           /* OUT - minimum depth number */
     mfu_flist** outlists /* OUT - array of lists split by depth */
 );
+#if 0
+int DCOPY_open_file(
+    const char* file,
+    int read,
+    mfu_copy_file_cache_t* cache
+);
+
+void DCOPY_copy_xattrs(
+    mfu_flist flist,
+    uint64_t index,
+    const char* dest_path
+);
+
+void DCOPY_copy_ownership(
+    mfu_flist flist,
+    uint64_t index,
+    const char* dest_path
+);
+
+void DCOPY_copy_permissions(
+    mfu_flist flist,
+    uint64_t index,
+    const char* dest_path
+);
+
+void DCOPY_copy_timestamps(
+    mfu_flist flist,
+    uint64_t index,
+    const char* dest_path
+);
+#endif
+/* called by single process upon detection of a problem */
+void DCOPY_abort(
+    int code
+) __attribute__((noreturn));
+
+/* called globally by all procs to exit */
+void DCOPY_exit(
+    int code
+);
+
+int mfu_copy_close_file(mfu_copy_file_cache_t* cache);
+
+void mfu_copy_set_metadata(int levels, int minlevel, mfu_flist* lists);
 
 /* create directories, we work from shallowest level to the deepest
  * with a barrier in between levels, so that we don't try to create
@@ -144,7 +296,7 @@ int mfu_create_directories(int levels, int minlevel, mfu_flist* lists);
 
 int mfu_create_files(int levels, int minlevel, mfu_flist* lists);
 
-void mfu_copy_files(mfu_flist list);
+void mfu_copy_files(mfu_flist list, uint64_t chunk_size);
 
 /* frees array of lists created in call to
  * mfu_flist_split_by_depth */
@@ -276,7 +428,6 @@ mfu_file_chunk* mfu_file_chunk_list_alloc(mfu_flist list, uint64_t chunk_size);
 
 /* free the linked list allocated with mfu_file_chunk_list_alloc */
 void mfu_file_chunk_list_free(mfu_file_chunk** phead);
-
 #endif /* MFU_FLIST_H */
 
 /* enable C++ codes to include this header directly */
