--- src/dcp/dcp.h
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2013-2015, Lawrence Livermore National Security, LLC.
- *   Produced at the Lawrence Livermore National Laboratory
- *   CODE-673838
- *
- * Copyright (c) 2006-2007,2011-2015, Los Alamos National Security, LLC.
- *   (LA-CC-06-077, LA-CC-10-066, LA-CC-14-046)
- *
- * Copyright (2013-2015) UT-Battelle, LLC under Contract No.
- *   DE-AC05-00OR22725 with the Department of Energy.
- *
- * Copyright (c) 2015, DataDirect Networks, Inc.
- * 
- * All rights reserved.
- *
- * This file is part of mpiFileUtils.
- * For details, see https://github.com/hpc/fileutils.
- * Please also read the LICENSE file.
-*/
-
-#ifndef __DCP_H_
-#define __DCP_H_
-
-#include "common.h"
-
-void DCOPY_print_usage(void);
-
-#endif /* __DCP_H_ */
